+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/AbstractWComponentTestCase.java
@@ -1,8 +1,10 @@
 package com.github.bordertech.wcomponents;
 
 import com.github.bordertech.wcomponents.util.SystemException;
+import java.io.Serializable;
 import java.lang.management.ManagementFactory;
 import java.lang.management.ThreadMXBean;
+import java.lang.reflect.Method;
 import java.util.Arrays;
 import junit.framework.Assert;
 import org.apache.commons.beanutils.PropertyUtils;
@@ -70,13 +72,43 @@ public abstract class AbstractWComponentTestCase {
 	protected void assertAccessorsCorrect(final WComponent component, final String method,
 			final Object initValue,
 			final Object defaultValue, final Object userContextValue) {
+
+		assertAccessorsCorrect(component, method, initValue, defaultValue, userContextValue, null);
+	}
+
+	/**
+	 * This method will test that the getter/setter methods on a component are returning the correct values in its (i)
+	 * initial state (ii) default state and (iii) user context.
+	 * <p>
+	 * If a setter method on the component has a variable argument, then an array of the variable argument type needs to
+	 * be provided via the setterArgs parameter.
+	 * </p>
+	 * <p>
+	 * For example, if the variable argument is a type {@link Serializable} then the setterArgs parameter would be set
+	 * as new Serializable[]{}. Values can also be passed in via the array.
+	 * </p>
+	 * <p>
+	 * Note that the component will be left in a dirty state after this method is invoked and the UIContext will be
+	 * reset.
+	 * </p>
+	 *
+	 * @param component the component to test the accessors on
+	 * @param method the method to test
+	 * @param initValue the initial value expected from the component
+	 * @param defaultValue the default value to be used on the shared model
+	 * @param userContextValue the value to be used with a user context
+	 * @param setterArgs array matching the variable argument type
+	 */
+	protected void assertAccessorsCorrect(final WComponent component, final String method,
+			final Object initValue,
+			final Object defaultValue, final Object userContextValue, final Object[] setterArgs) {
 		try {
 			// Check initial value
 			Object getvalue = invokeGetMethod(component, method);
 			checkValue(method, "Initial value.", initValue, getvalue);
 
 			// Set default value
-			invokeSetMethod(component, method, defaultValue);
+			invokeSetMethod(component, method, defaultValue, setterArgs);
 
 			// Check default value set correctly
 			getvalue = invokeGetMethod(component, method);
@@ -93,7 +125,7 @@ public abstract class AbstractWComponentTestCase {
 			checkValue(method, "User default value.", defaultValue, getvalue);
 
 			// Set user value
-			invokeSetMethod(component, method, userContextValue);
+			invokeSetMethod(component, method, userContextValue, setterArgs);
 
 			// Check user value
 			getvalue = invokeGetMethod(component, method);
@@ -116,6 +148,7 @@ public abstract class AbstractWComponentTestCase {
 	 * @return the value returned by the getter method
 	 */
 	private Object invokeGetMethod(final WComponent component, final String methodName) {
+		// Try property utils first
 		try {
 			return PropertyUtils.getProperty(component, methodName);
 		} catch (Exception e) {
@@ -123,17 +156,27 @@ public abstract class AbstractWComponentTestCase {
 					"Failed to get value on component for method " + methodName + " on "
 					+ component.getClass(), e);
 		}
+
 	}
 
 	/**
 	 * @param component the component to invoke the setter method on
 	 * @param methodName the name of the method
 	 * @param value the value to pass into the setter method
+	 * @param args if required the variable args
 	 */
 	private void invokeSetMethod(final WComponent component, final String methodName,
-			final Object value) {
+			final Object value, final Object[] args) {
 		try {
-			PropertyUtils.setProperty(component, methodName, value);
+			if (args == null) {
+				PropertyUtils.setProperty(component, methodName, value);
+			} else {
+				// Invoke specifying the variable arg type as propertyUtils cannot handle this
+				String setter = "set" + methodName.substring(0, 1).toUpperCase() + methodName.substring(1);
+				Class[] argTypes = new Class[]{value.getClass(), args.getClass()};
+				Method method = component.getClass().getMethod(setter, argTypes);
+				method.invoke(component, value, args);
+			}
 		} catch (Exception e) {
 			throw new SystemException(
 					"Failed to set value on component for method " + methodName + " on "